Extract context update method to `LocaleUtils`
