+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/BaseActivity.kt
@@ -1,12 +1,10 @@
 package ca.etsmtl.applets.etsmobile.presentation
 
 import android.content.Context
-import android.os.Build
 import android.os.Bundle
-import ca.etsmtl.applets.etsmobile.R
+import ca.etsmtl.applets.etsmobile.util.LocaleUtils
 import dagger.android.support.DaggerAppCompatActivity
 import model.UserCredentials
-import java.util.Locale
 
 /**
  * This base activity saves the user's credentials when [onSaveInstanceState] is called and restore
@@ -36,29 +34,6 @@ abstract class BaseActivity : DaggerAppCompatActivity() {
     }
 
     override fun attachBaseContext(newBase: Context) {
-        val context = newBase.createLanguagePrefConfigurationContext()
-
-        super.attachBaseContext(context)
-    }
-
-    @Suppress("DEPRECATION")
-    private fun Context.createLanguagePrefConfigurationContext(): Context {
-        val prefsFileName = getString(R.string.key_preference_file)
-        val prefs = getSharedPreferences(prefsFileName, Context.MODE_PRIVATE)
-        val langPrefKey = getString(R.string.key_language_pref)
-        val lang = prefs.getString(langPrefKey, null) ?: getString(R.string.default_entry_value_language_pref)
-        val locale = Locale(lang, "CA")
-        Locale.setDefault(locale)
-        val overrideConfiguration = resources
-            .configuration
-            .apply {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                    setLocale(locale)
-                } else {
-                    this.locale = locale
-                }
-            }
-
-        return createConfigurationContext(overrideConfiguration)
+        super.attachBaseContext(LocaleUtils.updateContext(newBase))
     }
 }
